Break up SnapshotState.kt into separate files.
SnapshotState.kt is a large file with many different public APIs that
are all technically related to snapshot state but aren't otherwise
related to each other. For example, derivedStateOf and its
implementation is a non-trivial chunk of code that makes sense to have
its own file, and makes it a little easier for readers to tell which
code is related to what.

Preserves JVM binary compatibility by merging all new files introduced by
this change with the @JvmMultifileClass feature.

Test: ./gradlew :compose:r:r:tDUT
Relnote: """Snapshot code was split up into multiple files, but all still
    lives in the same JVM class."""
